requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.15
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.0
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
flask-mongoengine-3==1.1.0
Flask-WTF==1.2.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
ordered-set==4.1.0
packaging==24.2
prompt_toolkit==3.0.48
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
rich==13.9.4
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.16.0
WTForms==3.2.1
zipp==3.21.0
